Writing/reintroducing tests
Summary: Brought back tests from the Bilby and Viterbi modules to cover some of the same ground in CWFollowup. Also updated the forms slightly, changing JobForm to NewJob and making it a Page. Added PageNav atom.

Test Plan: This was always the plan.

Reviewers: llakerink, mr-asher

Reviewed By: llakerink

Differential Revision: https://phab.adacs.org.au/D724

+++ src/cwfollowup/schema.py
@@ -123,7 +123,7 @@ class CWFollowupJobNode(DjangoObjectType):
 class CWFollowupResultFile(graphene.ObjectType):
     path = graphene.String()
     is_dir = graphene.Boolean()
-    file_size = graphene.Int()
+    file_size = graphene.Decimal()
     download_token = graphene.String()
